cloze validation + some fixes to the generator
